Switch from whoisd to local LDAP queries
Now that we no longer have Apache config files to read and whoisd
requires a magic password, there's no longer any advantage to having a
separate whois daemon.

This merges the contents of the latest whoisd from the scripts
repo. The code is Python 3-clean, but F30's python3-webpy is broken
and the same code needs to run on F20 so for now we're still using
F20/F29's python-webpy.